[TM-3359] add and draw polygon functionality
