+++ src/pages/site/[uuid]/tabs/Polygons.tsx
@@ -1,57 +1,23 @@
-import { Box, Flex, TableCell, TableRow, Text } from "@chakra-ui/react";
+import { Box, Flex, Text } from "@chakra-ui/react";
 import { useT } from "@transifex/react";
-import { Checkbox } from "@worldresources/wri-design-systems";
-import { FC, memo, ReactNode, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { FC, useCallback, useEffect, useMemo, useRef, useState } from "react";
 
 import PolygonsMap from "@/components/elements/Map-mapbox/components/PolygonsMap";
-import { OverlapPolygonPoint } from "@/components/elements/Map-mapbox/layers/overlapTypes";
 import PageContent from "@/components/extensive/PageElements/PageContent/PageContent";
 import PageItem from "@/components/extensive/PageElements/PageItem/PageItem";
+import LoadingContainer from "@/components/generic/Loading/LoadingContainer";
 import { useAllSitePolygons } from "@/connections/SitePolygons";
-import { useAllSiteValidations } from "@/connections/Validation";
-import { restorationStrategyType, targetLandUseType } from "@/constants/polygons";
+import { useMapAreaContext } from "@/context/mapArea.provider";
 import { SiteFullDto } from "@/generated/v3/entityService/entityServiceSchemas";
-import { SitePolygonsIndexQueryParams } from "@/generated/v3/researchService/researchServiceComponents";
 import { useDate } from "@/hooks/useDate";
 import { getThemedColor } from "@/lib/theme";
-import FeedbackTag from "@/redesignComponents/actions/Tags/FeedbackTag/FeedbackTag";
-import MappedTag, { MappedTagState } from "@/redesignComponents/actions/Tags/MappedTag/MappedTag";
-import ValidationTag, { ValidationTagState } from "@/redesignComponents/actions/Tags/ValidationTag/ValidationTag";
-import Tooltip from "@/redesignComponents/actions/Tooltip/Tooltip";
 import ResizeBox from "@/redesignComponents/containers/ResizableSplitView/ResizableBox";
 import MetricCard from "@/redesignComponents/dataDisplay/Metrics/MetricCard";
 import Table from "@/redesignComponents/dataDisplay/Table/Table";
 import { useTableSelection } from "@/redesignComponents/dataDisplay/Table/useTableSelection";
-import {
-  AgriculturalLandIcon,
-  AgroforestyIcon,
-  AreaHectaresIcon,
-  AssistedNaturalRegenIcon,
-  CalendarIcon,
-  DirectSeedingIcon,
-  DownloadIcon,
-  GrasslandIcon,
-  MangroveIcon,
-  NaturalForestIcon,
-  OpenNaturalEcosystemIcon,
-  PeatlandIcon,
-  PlusIcon,
-  SilvopastureIcon,
-  TreeIcon,
-  TreePlantingIcon,
-  UrbanForestIcon,
-  WetlandIcon,
-  WoodlotIcon
-} from "@/redesignComponents/foundations/Icons";
+import { AreaHectaresIcon, DownloadIcon, PlusIcon, TreeIcon } from "@/redesignComponents/foundations/Icons";
 import BulkActionToolbar from "@/redesignComponents/navigation/Toolbar/BulkActionToolbar";
-import { SelectedFilter } from "@/redesignComponents/navigation/Toolbar/ToolBar.type";
 import InlineMessage from "@/redesignComponents/status/InlineMessage/InlineMessage";
-import { OVERLAPPING_CRITERIA_ID } from "@/types/validation";
-import { formatNumberLocaleString } from "@/utils/dashboardUtils";
-import {
-  mapSitePolygonStatusToMappedTagState,
-  mapSitePolygonValidationStatusToValidationTagState
-} from "@/utils/mapStatusToTagStateEntity";
 
 import DeletePolygon from "../components/Modals/DeletePolygon";
 import MatchingPolygonsFound from "../components/Modals/MatchingPolygonsFound";
@@ -60,231 +26,24 @@ import SubmitPolygons from "../components/Modals/SubmitPolygons";
 import UploadError from "../components/Modals/UploadError";
 import UploadPhotos from "../components/Modals/UploadPhotos";
 import UploadPolygons from "../components/Modals/UploadPolygons";
-import {
-  EMPTY_POLYGON_FILTERS,
-  PolygonFilterState,
-  RESTORATION_PRACTICE_LABELS,
-  SUBMISSION_STATUS_LABELS,
-  TARGET_LAND_USE_LABELS,
-  VALIDATION_STATUS_LABELS
-} from "../components/polygonFilter.constants";
+import { PolygonRow, PolygonTableRow } from "../components/PolygonTableRow";
 import PolygonToolbar from "../components/PolygonToolbar";
+import { useSitePolygonFilters } from "../hooks/useSitePolygonFilters";
+import { useSitePolygonOverlap } from "../hooks/useSitePolygonOverlap";
+import { useSitePolygonTableData } from "../hooks/useSitePolygonTableData";
+import { useStartSitePolygonDrawing } from "../hooks/useStartSitePolygonDrawing";
 
 interface SitePolygonsTabProps {
   site: SiteFullDto;
 }
 
-type SiteTypeConfig = { icon: ReactNode; label: string; tooltip?: string };
-
-export type PolygonTableRow = {
-  id: string;
-  polygonName: string;
-  submission: MappedTagState;
-  validation: ValidationTagState;
-  restorationPractice: restorationStrategyType[];
-  targetLandUse: targetLandUseType | null;
-  plantingDate: string;
-  treeDistribution: string[];
-  treesPlanted: number;
-  area: number;
-};
-
-const TARGET_LAND_USE_MAP: Record<targetLandUseType, SiteTypeConfig> = {
-  agroforest: {
-    icon: <AgroforestyIcon boxSize={3.5} />,
-    label: "Agroforest"
-  },
-  "agricultural-land": {
-    icon: <AgriculturalLandIcon boxSize={3.5} />,
-    label: "Agricultural Land"
-  },
-  grassland: {
-    icon: <GrasslandIcon boxSize={3.5} />,
-    label: "Grassland"
-  },
-  mangrove: {
-    icon: <MangroveIcon boxSize={3.5} />,
-    label: "Mangrove"
-  },
-  "open-natural-ecosystem": {
-    icon: <OpenNaturalEcosystemIcon boxSize={3.5} />,
-    label: "Open Natural Ecosystem"
-  },
-  "natural-forest": {
-    icon: <NaturalForestIcon boxSize={3.5} />,
-    label: "Natural Forest"
-  },
-  peatland: {
-    icon: <PeatlandIcon boxSize={3.5} />,
-    label: "Peatland"
-  },
-  "riparian-area-or-wetland": {
-    icon: <WetlandIcon boxSize={3.5} />,
-    label: "Riparian Area / Wetland"
-  },
-  silvopasture: {
-    icon: <SilvopastureIcon boxSize={3.5} />,
-    label: "Silvopasture"
-  },
-  "urban-forest": {
-    icon: <UrbanForestIcon boxSize={3.5} />,
-    label: "Urban Forest"
-  },
-  "woodlot-or-plantation": {
-    icon: <WoodlotIcon boxSize={3.5} />,
-    label: "Woodlot / Plantation"
-  }
-};
-
-const SITE_RESTORATION_STRATEGY_MAP: Record<restorationStrategyType, ReactNode> = {
-  "tree-planting": (
-    <Tooltip content="Tree planting">
-      <TreePlantingIcon boxSize={5} color="secondary.800" />
-    </Tooltip>
-  ),
-  "assisted-natural-regeneration": (
-    <Tooltip content="Assisted natural regeneration (ANR)">
-      <AssistedNaturalRegenIcon boxSize={5} color="secondary.800" />
-    </Tooltip>
-  ),
-  "direct-seeding": (
-    <Tooltip content="Direct seeding">
-      <DirectSeedingIcon boxSize={5} color="secondary.800" />
-    </Tooltip>
-  )
-};
-
-const isRestorationStrategy = (value: string): value is restorationStrategyType => {
-  return value === "tree-planting" || value === "assisted-natural-regeneration" || value === "direct-seeding";
-};
-
-const isTargetLandUseType = (value: string): value is targetLandUseType => {
-  return value in TARGET_LAND_USE_MAP;
-};
-
-const formatDistributionValue = (value: string): string => {
-  return value
-    .split("-")
-    .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(" ");
-};
-
-const getTargetLandUseConfig = (targetLandUse: targetLandUseType | null) => {
-  if (targetLandUse == null) {
-    return { icon: null, label: "-" };
-  }
-  return TARGET_LAND_USE_MAP[targetLandUse];
-};
-
-const renderRestorationPractice = (restorationPractice: restorationStrategyType[]) => {
-  if (restorationPractice.length === 0) {
-    return <Text>—</Text>;
-  }
-
-  return (
-    <Flex className="items-center gap-2">
-      {restorationPractice.map((rp, index) => (
-        <Flex key={`${rp}-${index}`} className="items-center gap-2">
-          {SITE_RESTORATION_STRATEGY_MAP[rp]}
-          {index < restorationPractice.length - 1 && <PlusIcon boxSize={2.5} color="secondary.800" />}
-        </Flex>
-      ))}
-    </Flex>
-  );
-};
-
-const HOVERED_ROW_STYLE: React.CSSProperties = {
-  backgroundColor: getThemedColor("primary", 100),
-  borderBottom: `2px solid ${getThemedColor("primary", 700)}`
-};
-
-interface PolygonRowProps {
-  row: PolygonTableRow;
-  rowProps?: Record<string, unknown>;
-  isSelected: boolean;
-  isHovered: boolean;
-  onHover: (uuid: string) => void;
-  onSelectChange: (row: PolygonTableRow, checked: boolean) => void;
-}
-
-const PolygonRowComponent: FC<PolygonRowProps> = ({
-  row,
-  rowProps,
-  isSelected,
-  isHovered,
-  onHover,
-  onSelectChange
-}) => {
-  const handleOnRowSelected = useCallback(
-    ({ checked }: { checked?: boolean | "indeterminate" }) => {
-      onSelectChange(row, Boolean(checked));
-    },
-    [row, onSelectChange]
-  );
-
-  const handleMouseEnter = useCallback(() => {
-    onHover(row.id);
-  }, [row.id, onHover]);
-
-  const targetLandUseConfig = getTargetLandUseConfig(row.targetLandUse);
-  return (
-    <TableRow
-      {...(rowProps as Record<string, unknown>)}
-      aria-selected={isSelected || isHovered}
-      onMouseEnter={handleMouseEnter}
-      style={isHovered ? HOVERED_ROW_STYLE : undefined}
-    >
-      <TableCell>
-        <Checkbox name={`checkbox-${row.id}`} onCheckedChange={handleOnRowSelected} checked={isSelected} />
-      </TableCell>
-      <TableCell className="min-w-[17.75rem] max-w-[17.75rem]">
-        <Box>
-          <Text textStyle="400-bold" color="neutral.800" className="truncate">
-            {row.polygonName ?? "—"}
-          </Text>
-        </Box>
-      </TableCell>
-      <TableCell className="min-w-[15.875rem]">
-        {row.submission != null ? <MappedTag state={row.submission} /> : <Text>—</Text>}
-      </TableCell>
-      <TableCell className="min-w-[12.75rem]">
-        {row.validation != null ? <ValidationTag status={row.validation} /> : <Text>—</Text>}
-      </TableCell>
-      <TableCell className="min-w-[15.5rem]">
-        <Flex className="items-center gap-2">{renderRestorationPractice(row.restorationPractice)}</Flex>
-      </TableCell>
-      <TableCell className="min-w-[16.75rem]">
-        {targetLandUseConfig.icon != null && targetLandUseConfig.label != null ? (
-          <Flex className="items-center gap-2" color="neutral.800">
-            {targetLandUseConfig.icon}
-            <Text>{targetLandUseConfig.label}</Text>
-          </Flex>
-        ) : (
-          <Text>—</Text>
-        )}
-      </TableCell>
-      <TableCell className="min-w-[11.5rem]">
-        <FeedbackTag
-          type="info-grey"
-          className="w-fit"
-          label={row.plantingDate != "-" ? row.plantingDate : "—"}
-          icon={<CalendarIcon boxSize={2.5} />}
-        />
-      </TableCell>
-      <TableCell className="min-w-[15.875rem]">
-        <Text>{row.treeDistribution.length > 0 ? row.treeDistribution.join(", ") : "—"}</Text>
-      </TableCell>
-      <TableCell className="min-w-[12.75rem]">{formatNumberLocaleString(row.treesPlanted) ?? "—"}</TableCell>
-      <TableCell className="min-w-[15.75rem]">{formatNumberLocaleString(row.area) ?? "—"}</TableCell>
-    </TableRow>
-  );
-};
-
-const PolygonRow = memo(PolygonRowComponent);
+export type { PolygonTableRow } from "../components/PolygonTableRow";
 
 const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
   const t = useT();
   const { format } = useDate();
+  const { setSiteData, resetSiteMapInteractionState } = useMapAreaContext();
+
   const tableContainerRef = useRef<HTMLDivElement>(null);
   const [showUploadModal, setShowUploadModal] = useState(false);
   const [showMatchingPolygonsFoundModal, setMatchingPolygonsFoundModal] = useState(false);
@@ -292,111 +51,55 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
   const [showSubmitPolygonsModal, setSubmitPolygonsModal] = useState(false);
   const [showDeletePolygonModal, setDeletePolygonModal] = useState(false);
   const [showUploadErrorModal, setUploadErrorModal] = useState(false);
-  const [showUploadPhotosModal, setUploadPhotosModal] = useState(false);
-
+  const [showUploadPhotosModal, setShowUploadPhotosModal] = useState(false);
   const [isStickyActive, setIsStickyActive] = useState(false);
-  const [polygonSearch, setPolygonSearch] = useState("");
-  const [debouncedPolygonSearch, setDebouncedPolygonSearch] = useState("");
-  const [polygonFilters, setPolygonFilters] = useState<PolygonFilterState>(EMPTY_POLYGON_FILTERS);
+  const [hoveredPolygonUuid, setHoveredPolygonUuid] = useState<string | null>(null);
 
-  useEffect(() => {
-    const timeout = window.setTimeout(() => {
-      setDebouncedPolygonSearch(polygonSearch.trim());
-    }, 300);
+  const {
+    polygonSearch,
+    polygonFilters,
+    sitePolygonFilter,
+    activeFilterLabels,
+    setPolygonSearch,
+    setPolygonFilters,
+    handleClearPolygonFilters
+  } = useSitePolygonFilters({ t });
 
-    return () => window.clearTimeout(timeout);
-  }, [polygonSearch]);
-
-  const sitePolygonFilter = useMemo(() => {
-    const filter: Record<string, unknown> = {};
-    if (polygonFilters.polygonStatus.length > 0) filter.polygonStatus = polygonFilters.polygonStatus;
-    if (polygonFilters.validationStatus.length > 0) filter.validationStatus = polygonFilters.validationStatus;
-    if (polygonFilters.plantStartFrom !== "") filter.plantStartFrom = `${polygonFilters.plantStartFrom}T00:00:00.000Z`;
-    if (polygonFilters.plantStartTo !== "") filter.plantStartTo = `${polygonFilters.plantStartTo}T00:00:00.000Z`;
-    if (polygonFilters.practice.length > 0) filter.practice = polygonFilters.practice;
-    if (polygonFilters.targetSys.length > 0) filter.targetSys = polygonFilters.targetSys;
-    if (polygonFilters.hasOverlap) filter.hasOverlap = true;
-    if (debouncedPolygonSearch !== "") {
-      filter.search = debouncedPolygonSearch;
-      filter.searchFields = ["polyName", "polygonUuid"];
-    }
-    return filter as Partial<SitePolygonsIndexQueryParams>;
-  }, [debouncedPolygonSearch, polygonFilters]);
-
-  const { data: polygonsData = [], refetch: refetchPolygons } = useAllSitePolygons({
+  const {
+    data: polygonsData = [],
+    isLoading: isLoadingPolygons,
+    error: polygonLoadError,
+    progress: polygonLoadProgress,
+    total: polygonLoadTotal,
+    refetch: refetchPolygons
+  } = useAllSitePolygons({
     entityName: "sites",
     entityUuid: site.uuid,
     enabled: site.uuid != null && site.uuid !== "",
     filter: sitePolygonFilter
   });
 
-  const { allValidations: overlapValidations, fetchAllValidationPages: fetchOverlapValidations } =
-    useAllSiteValidations(site.uuid, OVERLAPPING_CRITERIA_ID);
-
-  const polygonIdsKey = useMemo(
-    () =>
-      polygonsData
-        .map(p => p.uuid)
-        .filter((uuid): uuid is string => uuid != null && uuid !== "")
-        .sort()
-        .join(","),
-    [polygonsData]
-  );
-
-  useEffect(() => {
-    if (site.uuid == null || site.uuid === "") {
-      return;
-    }
-    void fetchOverlapValidations();
-  }, [site.uuid, polygonIdsKey, fetchOverlapValidations]);
-
-  const { polygonsWithOverlapCount, overlapPolygons } = useMemo(() => {
-    const currentPolygonUuids = new Set(
-      polygonsData
-        .map(polygon => polygon.polygonUuid ?? polygon.uuid)
-        .filter((id): id is string => id != null && id !== "")
-    );
-    const overlapPolygonUuids = new Set(
-      overlapValidations
-        .map(v => v.polygonUuid)
-        .filter((id): id is string => id != null && id !== "" && currentPolygonUuids.has(id))
-    );
-    if (overlapPolygonUuids.size === 0) {
-      return { polygonsWithOverlapCount: 0, overlapPolygons: [] as OverlapPolygonPoint[] };
-    }
-    const points: OverlapPolygonPoint[] = [];
-    for (const polygon of polygonsData) {
-      const uuid = polygon.polygonUuid ?? polygon.uuid;
-      if (uuid == null || !overlapPolygonUuids.has(uuid)) continue;
-      if (polygon.lat == null || polygon.long == null) continue;
-      points.push({ polygonUuid: uuid, lat: polygon.lat, lng: polygon.long });
-    }
-    return { polygonsWithOverlapCount: points.length, overlapPolygons: points };
-  }, [overlapValidations, polygonsData]);
-
-  const polygonRows = useMemo<PolygonTableRow[]>(
-    () =>
-      polygonsData.map(polygon => ({
-        id: polygon.polygonUuid ?? polygon.uuid,
-        polygonName: polygon.name ?? t("Unnamed Polygon"),
-        submission: mapSitePolygonStatusToMappedTagState(polygon.status),
-        validation: mapSitePolygonValidationStatusToValidationTagState(polygon.validationStatus),
-        restorationPractice: (polygon.practice ?? []).filter(isRestorationStrategy),
-        targetLandUse: polygon.targetSys != null && isTargetLandUseType(polygon.targetSys) ? polygon.targetSys : null,
-        plantingDate: polygon.plantStart != null ? format(polygon.plantStart, "yyyy-MM-dd") : "-",
-        treeDistribution: (polygon.distr ?? []).map(formatDistributionValue),
-        treesPlanted: polygon.numTrees ?? 0,
-        area: polygon.calcArea ?? 0
-      })),
-    [format, polygonsData, t]
-  );
+  const { polygonRows, columns, totalTreesPlanted, totalRestorationAreaHa } = useSitePolygonTableData({
+    polygonsData,
+    t,
+    format
+  });
+  const { polygonsWithOverlapCount, overlapPolygons } = useSitePolygonOverlap({ siteUuid: site.uuid, polygonsData });
 
   const { selectedRows, selectedRowIds, setSelectedRowIds, handleRowSelected, onAllItemsSelected } =
     useTableSelection<PolygonTableRow>(true, polygonRows);
 
-  const [hoveredPolygonUuid, setHoveredPolygonUuid] = useState<string | null>(null);
-
   const selectedPolygonUuids = useMemo(() => Array.from(selectedRowIds, id => String(id)), [selectedRowIds]);
+
+  useEffect(() => {
+    setSiteData(site);
+  }, [setSiteData, site]);
+
+  useEffect(() => {
+    return () => {
+      resetSiteMapInteractionState();
+    };
+  }, [resetSiteMapInteractionState]);
 
   useEffect(() => {
     const visibleRowIds = new Set(polygonRows.map(row => row.id));
@@ -405,6 +108,12 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
       return next.size === prev.size ? prev : next;
     });
   }, [polygonRows, setSelectedRowIds]);
+
+  const clearTableSelection = useCallback(() => {
+    setSelectedRowIds(new Set<string>());
+  }, [setSelectedRowIds]);
+
+  const startDrawing = useStartSitePolygonDrawing({ onClearTableSelection: clearTableSelection });
 
   const handlePolygonClickedFromMap = useCallback(
     (uuid: string) => {
@@ -428,83 +137,9 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
     [hoveredPolygonUuid, selectedPolygonUuids, handlePolygonClickedFromMap]
   );
 
-  const handleClearHover = useCallback(() => setHoveredPolygonUuid(null), []);
-
-  const handleClearPolygonFilters = useCallback(() => {
-    setPolygonFilters(EMPTY_POLYGON_FILTERS);
-    setPolygonSearch("");
-    setDebouncedPolygonSearch("");
+  const handleClearHover = useCallback(() => {
+    setHoveredPolygonUuid(null);
   }, []);
-
-  const activeFilterLabels = useMemo<SelectedFilter[]>(() => {
-    const labels: SelectedFilter[] = [];
-    if (polygonFilters.polygonStatus.length > 0) {
-      labels.push({
-        label: polygonFilters.polygonStatus.map(status => t(SUBMISSION_STATUS_LABELS[status])),
-        onRemove: () => {
-          setPolygonFilters(current => ({ ...current, polygonStatus: [] }));
-        },
-        category: t("Submission")
-      });
-    }
-    if (polygonFilters.validationStatus.length > 0) {
-      labels.push({
-        label: polygonFilters.validationStatus.map(status => t(VALIDATION_STATUS_LABELS[status])),
-        onRemove: () => {
-          setPolygonFilters(current => ({ ...current, validationStatus: [] }));
-        },
-        category: t("Validation")
-      });
-    }
-    if (polygonFilters.plantStartFrom !== "" || polygonFilters.plantStartTo !== "") {
-      const fromLabel = polygonFilters.plantStartFrom !== "" ? polygonFilters.plantStartFrom : t("Any date");
-      const toLabel = polygonFilters.plantStartTo !== "" ? polygonFilters.plantStartTo : t("Any date");
-      labels.push({
-        label: `${fromLabel} - ${toLabel}`,
-        onRemove: () => {
-          setPolygonFilters(current => ({ ...current, plantStartFrom: "", plantStartTo: "" }));
-        }
-      });
-    }
-    if (polygonFilters.practice.length > 0) {
-      labels.push({
-        label: polygonFilters.practice.map(practice => t(RESTORATION_PRACTICE_LABELS[practice])),
-        onRemove: () => {
-          setPolygonFilters(current => ({ ...current, practice: [] }));
-        }
-      });
-    }
-    if (polygonFilters.targetSys.length > 0) {
-      labels.push({
-        label: polygonFilters.targetSys.map(targetSys => t(TARGET_LAND_USE_LABELS[targetSys])),
-        onRemove: () => {
-          setPolygonFilters(current => ({ ...current, targetSys: [] }));
-        }
-      });
-    }
-    if (polygonFilters.hasOverlap) {
-      labels.push({
-        label: t("Failed overlap validation"),
-        onRemove: () => {
-          setPolygonFilters(current => ({ ...current, hasOverlap: false }));
-        }
-      });
-    }
-    return labels;
-  }, [polygonFilters, t, setPolygonFilters]);
-
-  const { totalTreesPlanted, totalRestorationAreaHa } = useMemo(() => {
-    let trees = 0;
-    let area = 0;
-    for (const polygon of polygonsData) {
-      trees += polygon.numTrees ?? 0;
-      area += polygon.calcArea ?? 0;
-    }
-    return {
-      totalTreesPlanted: trees,
-      totalRestorationAreaHa: Math.round(area * 100) / 100
-    };
-  }, [polygonsData]);
 
   const { selectedTreesPlanted, selectedRestorationAreaHa } = useMemo(
     () =>
@@ -520,56 +155,7 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
 
   const selectedRestorationAreaRounded = Math.round(selectedRestorationAreaHa * 100) / 100;
   const hasPolygonSelection = selectedRows.length > 0;
-
-  const columns = useMemo(
-    () => [
-      {
-        key: "polygonName",
-        label: t("Polygon Name"),
-        sortable: true
-      },
-      {
-        key: "submission",
-        label: t("Submission"),
-        sortable: true
-      },
-      {
-        key: "validation",
-        label: t("Validation"),
-        sortable: true
-      },
-      {
-        key: "restorationPractice",
-        label: t("Restoration Practice")
-      },
-      {
-        key: "targetLandUse",
-        label: t("Target Land Use"),
-        sortable: true
-      },
-      {
-        key: "plantingDate",
-        label: t("Planting Date"),
-        sortable: true
-      },
-      {
-        key: "treeDistribution",
-        label: t("Tree Distribution"),
-        sortable: true
-      },
-      {
-        key: "treesPlanted",
-        label: t("Trees Planted"),
-        sortable: true
-      },
-      {
-        key: "area",
-        label: t("Area (ha)"),
-        sortable: true
-      }
-    ],
-    [t]
-  );
+  const shouldShowNoResults = !isLoadingPolygons && polygonRows.length === 0;
 
   const selectableRenderRow = useCallback(
     (row: PolygonTableRow, rowProps?: Record<string, unknown>) => (
@@ -585,10 +171,8 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
     [handleRowSelected, hoveredPolygonUuid, selectedRowIds]
   );
 
-  const getPolygonsTableStyles = (isStickyActive: boolean) => ({
-    "& table td": {
-      height: "3rem"
-    },
+  const getPolygonsTableStyles = (isStickyTableActive: boolean) => ({
+    "& table td": { height: "3rem" },
     "& table th:first-child": {
       position: "sticky",
       left: 0,
@@ -627,7 +211,7 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
       display: "flex",
       alignItems: "center",
       height: "100%",
-      ...(isStickyActive && {
+      ...(isStickyTableActive && {
         borderRight: `1px solid ${getThemedColor("neutral", 400)}`
       })
     }
@@ -642,6 +226,11 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
     container.addEventListener("scroll", handleScroll);
     return () => container.removeEventListener("scroll", handleScroll);
   }, []);
+
+  const loadingLabel =
+    polygonLoadTotal > 0
+      ? t("Loading polygons ({loaded}/{total})", { loaded: polygonLoadProgress, total: polygonLoadTotal })
+      : t("Loading polygons");
 
   return (
     <PageContent className="bg-theme-neutral-100">
@@ -658,19 +247,15 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
           mainActionLabel: t("Add"),
           size: "small",
           leftIcon: <PlusIcon />,
-          mainActionOnClick: () => {},
+          mainActionOnClick: startDrawing,
           otherActions: [
-            {
-              label: t("Draw Polygon"),
-              onClick: () => {},
-              value: "draft"
-            },
+            { label: t("Draw Polygon"), onClick: startDrawing, value: "draw-polygon" },
             {
               label: t("Upload"),
               onClick: () => {
                 setShowUploadModal(true);
               },
-              value: "save-close"
+              value: "upload"
             }
           ],
           variant: "primary"
@@ -686,30 +271,23 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
           onClearFilters={handleClearPolygonFilters}
         />
       </PageItem>
-      {selectedRows != null && selectedRows.length > 0 && (
-        <Box position={"fixed"} zIndex={"999999"} bottom={0} left={3} right={3}>
+
+      {selectedRows.length > 0 && (
+        <Box position="fixed" zIndex="999999" bottom={0} left={3} right={3}>
           <BulkActionToolbar
-            ButtonCancel={{
-              children: "Cancel"
-            }}
+            ButtonCancel={{ children: t("Cancel") }}
             ButtonDelete={{
-              children: "Delete",
+              children: t("Delete"),
               onClick: () => {
                 setDeletePolygonModal(true);
               }
             }}
-            items={selectedRows.length + ""}
-            primaryButtonProps={{
-              children: "Download"
-            }}
-            quantityButtonProps={{
-              children: "Run Validation"
-            }}
-            secondaryButtonProps={{
-              children: "Edit Details"
-            }}
+            items={String(selectedRows.length)}
+            primaryButtonProps={{ children: t("Download") }}
+            quantityButtonProps={{ children: t("Run Validation") }}
+            secondaryButtonProps={{ children: t("Edit Details") }}
             tertiaryButtonProps={{
-              children: "Submit",
+              children: t("Submit"),
               onClick: () => {
                 setSubmitPolygonsModal(true);
               }
@@ -717,22 +295,18 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
           />
         </Box>
       )}
+
       <UploadPolygons open={showUploadModal} onOpenChange={setShowUploadModal} />
       <MatchingPolygonsFound open={showMatchingPolygonsFoundModal} onOpenChange={setMatchingPolygonsFoundModal} />
       <PolygonSubmitted
         open={showPolygonSubmittedModal}
         onOpenChange={setPolygonSubmittedModal}
-        polygons={"Polygon Name A"}
-      />
-      <PolygonSubmitted
-        open={showPolygonSubmittedModal}
-        onOpenChange={setPolygonSubmittedModal}
-        polygons={["Polygon Name A", "polygon Name B"]}
+        polygons={["Polygon Name A", "Polygon Name B"]}
       />
       <SubmitPolygons open={showSubmitPolygonsModal} onOpenChange={setSubmitPolygonsModal} />
       <DeletePolygon open={showDeletePolygonModal} onOpenChange={setDeletePolygonModal} polygons={selectedRows} />
       <UploadError open={showUploadErrorModal} onOpenChange={setUploadErrorModal} />
-      <UploadPhotos open={showUploadPhotosModal} onOpenChange={setUploadPhotosModal} />
+      <UploadPhotos open={showUploadPhotosModal} onOpenChange={setShowUploadPhotosModal} />
 
       <ResizeBox initialHeight={100} minHeight={100} maxHeight={600}>
         <PolygonsMap
@@ -745,10 +319,33 @@ const SitePolygonsTab: FC<SitePolygonsTabProps> = ({ site }) => {
           overlapPolygons={overlapPolygons}
         />
       </ResizeBox>
-      {polygonRows?.length === 0 ? (
+
+      {polygonLoadError != null && (
+        <InlineMessage
+          className="mt-4"
+          variant="error"
+          label={t("Unable to load polygons")}
+          caption={t("Please retry loading polygons.")}
+          actionLabel={t("Retry")}
+          onActionClick={() => {
+            void refetchPolygons();
+          }}
+        />
+      )}
+
+      {isLoadingPolygons ? (
+        <Box className="py-4">
+          <LoadingContainer loading />
+          <Text textStyle="400" color="neutral.700" className="mt-2">
+            {loadingLabel}
+          </Text>
+        </Box>
+      ) : shouldShowNoResults ? (
         <Box>
-          <Text textStyle={"400-bold"}>No results found</Text>
-          <Text textStyle={"400"}>We couldn’t find any site areas matching your search. Try a different keyword.</Text>
+          <Text textStyle="400-bold">{t("No results found")}</Text>
+          <Text textStyle="400">
+            {t("We couldn’t find any site areas matching your search. Try a different keyword.")}
+          </Text>
         </Box>
       ) : (
         <>
